requirements.txt
bcrypt==3.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
dataclasses==0.8
importlib-metadata==4.8.3
isort==5.10.1
mypy-extensions==0.4.3
paramiko==2.9.2
pathspec==0.9.0
platformdirs==2.4.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.19.2
six==1.16.0
tomli==1.2.3
typed-ast==1.5.1
typing_extensions==4.0.1
zipp==3.6.0
